add common.h
