+++ common.h
@@ -1,0 +1,38 @@
+#ifndef COMMON_H
+#define COMMON_H
+
+#define MIN(a, b) a < b ? a : b
+
+static void die(char *s)
+{
+    fprintf(stderr, "%s", s);
+    exit(1);
+}
+
+static bool read_all(int fd, void *buf, size_t n)
+{
+    while (n > 0) {
+        ssize_t rv = read(fd, buf, n);
+        if (rv <= 0) {
+            return false;
+        }
+        n -= rv;
+        buf += rv;
+    }
+    return true;
+}
+
+static bool write_all(int fd, void *buf, size_t n)
+{
+    while (n > 0) {
+        ssize_t rv = write(fd, buf, n);
+        if (rv <= 0) {
+            return false;
+        }
+        n -= rv;
+        buf += rv;
+    }
+    return true;
+}
+
+#endif
